[FIX] l10n_ec_edi: fix withhold details in edi xml
Steps to reproduce:
- Install l10n_ec_edi
- Switch to an Ecuadorian company (e.g. EC Company)
- Create a vendor bill:
  * Vendor: [a contact not in Ecuador] (e.g. Azure Interieur)
  * Journal: 001-001 Liquidaciones de Compra
  * Document Type: (03) Liquidación de compra de Bienes o Prestación de servicios
  * Payment Method (SRI): [any] (e.g. Sin utilización del sistema financiero)
  * Invoice Lines => Product: [any] - Taxes: [an EC tax] (e.g. IVA 12% (510, 01 Crédito IVA))
- Confirm invoice
- Click on "ADD WITHHOLD"
- Add a withhold line with a tax having its l10n_ec_code_base value in L10N_EC_WTH_FOREIGN_GENERAL_REGIME_CODES list
(e.g. 501-422 22% Pago al Exterior - Beneficios Empresariales (Sin Convenio de Doble Tributación)
- Create and post Withhold
- Check the EDI xml linked to the withhold
=> <paisEfecPago/> element has a "NA" value, but it should be the l10n_ec_code_ats value
of the vendor country as we have used a tax having its l10n_ec_code_base code in
L10N_EC_WTH_FOREIGN_GENERAL_REGIME_CODES list

Cause:
Only the first tax from the withhold lines is checked to determine if it is
in L10N_EC_WTH_FOREIGN_GENERAL_REGIME_CODES list or not.
So, if the first tax is not in the list, the system considers that all taxes
are not in the list (even if the second one is in the list) and it doesn't
populate the correct data for <paisEfecPago/> element in the edi xml.
(same issue for <aplicConvDobTrib/> and <pagExtSujRetNorLeg/> elements)

Solution:
Check all the taxes instead of just the first one.

opw-3478597

closes odoo/enterprise#48463

Signed-off-by: Josse Colpaert <jco@odoo.com>