[FIX] l10n_be_hr_contract_salary : show canteen cost in details
Issue:
======
When an employee have canteen cost it doesnt show in the details of net
salary.

Steps to reproduce the error :
==============================
- Install l10n_be_hr_contract_salary
- Select My belgian company
- Go to employees and select Laurie
- Select in Contract Since smart button
- Go to salary information and add a canteen cost
- Click on generate offer and click save
- Go to the created link in the offer form.
- Click on details on net
- There is no canteen cost

Solution:
=========
In the override of the function , l10n_be_canteen _cost is always set to
False , updating it with the value in the contract solved the problem.

opw-3536125

closes odoo/enterprise#48565

X-original-commit: c59d8e96bc62341f6088b91a999ba84a9c5061e7
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>