[FIX] mrp_workorder_hr: don’t link employee to the workorder
Steps to reproduce the bug:
- Create 2 work centers, 1 requires log in the other one doesn't

- Create a storble product “P1”:
    - With BoM:
        - Operation:
            - OP1: Work center 1 with login
            - OP2: Work center 2 without login

- Create a MO with P1 and confirm it
- Enter the work order of the work center that requires log in, log in,
leave work order:
     - timer stops

- Now enter the work order that does not require log in, and leave work
order

Probelm:
"Timer does not stop and continues running.

When we open the tablet, we check if we should start the timer by
creating an 'mrp.workcenter.productivity' record linked to the work
order. In this case, we create one:

https://github.com/odoo/odoo/blob/67774c967b587a660be032712b7a2369ed7abc3a/addons/mrp/models/mrp_workorder.py#L613-L614

However, since we logged in with an employee in the first work order,
that employee remains logged in, So a second timer is created when we
try to link the employee to this work order:
https://github.com/odoo/enterprise/blob/d4f65146dacf0c9c27577ce0965fe1589d548171/mrp_workorder_hr/models/mrp_workorder.py#L46

Solution:
This work order uses a work center that does not require an employee
to be connected, so we should not link the connected employee to this
work order."

opw-3497588

closes odoo/enterprise#48482

Signed-off-by: William Henrotin (whe) <whe@odoo.com>