[FIX] documents: test with boolean toggle field
The changes introduced to BooleanToggleField by the PR: https://github.com/odoo/odoo/pull/136924,
mean that an update executed on a boolean field from a Component other
than the BooleanToggleField displaying it requires waiting for one more nextTick
for the change to be displayed. This is caused by useObserveRecord waiting
for an animationFrame to execute its callBack and therefore in this case
to apply the change of state to the BooleanToggleField.

closes odoo/enterprise#48515

Related: odoo/odoo#136924
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>