[FIX] l10n_lt_reports: Wrong hierarchy level in reports
Description of the issue this commit addresses:

When inside the Lithuanian BS and P&L reports, the first lines of the section
are not higlighted as the first one of the group. This is due to their
hierarchy level being set at the default value of one.

Also, when arriving for the first time on the report, only the hierarchy_level
0 lines are shown due to the reports being folded by default and thoses lines
having the attribute "foldable"

Desired behavior after this commit is merged:

This commit sets the hierarchy level of the wanted lines to 0 so they
get all the styling title lines normally have. Like the entire line
being on a gray back instead of white.

The commit also removes the foldable attribute from the hierarchy_level
0 lines of the BS report so that they are unfolded by default.

This commit makes it clearer for which lines are the first one
of the sections in the Lithuanian BS and P&L and makes the main lines
unfolded by default.

closes odoo/enterprise#48636

X-original-commit: 96d4a3cdf4e27fb095772be5eb53adbae81cbe4c
Related: odoo/odoo#138060
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>