[FIX] web_studio: report editor qweb render with studio in context
Before this commit when the report editor asked for the Qweb, the specific
rendering context did not contain the studio key.
It could produces issues down the line, as some reports are prepared to receive
that key and produce another output.

After this commit, the studio key is present in the context and in the data passed to render
the report.

closes odoo/enterprise#48538

X-original-commit: fd35cee117a88e817d19ae52a2c0c080cca5a805
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>