[FIX] account_asset: make gross_increase on remainder of life left
When we make a reevaluation with an increase, the date where you
make the increase has no impact.
It's depreciated for the whole duration of the parent asset.
It does not make sense, and should be from the moment where the
asset is reevaluated.

The child asset will have a lifetime depending on the time left
for the parent.
For degressive then linear, we try to follow the curve of the parent

opw-3526524
Part of pad of t-3540377

closes odoo/enterprise#50338

X-original-commit: 0f0970fb6157da8489cb361d401338e74a2c16ec
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>