[FIX] pos_restaurant_preparation_display: load data on new pos config
In this commit 2 issues are fixed. The first issue is when trying to load
demo data when the default pos configs are deleted. The second is when trying to
reload onboarding data when the products are already loaded but archived.

steps to reproduce the first issue:
- load db without demo data
- delete pos created by default
- create a new pos config
- open and try to load demo data

steps to reproduce the second issue:
- load db without demo data
- open pos and load demo data
- archive products
- open pos and try to load demo data

changes:
- adapt pos_restaurant_preparation_display to the separation of demo data
introduced in the community branch.

closes odoo/enterprise#48621

Related: odoo/odoo#138035
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>