[FIX] web_studio: tests: click helper: throw error if target is disabled
This commit is a backport of odoo/enterprise@3688e77d5482bd647ce4a304aefb846eab2f8794.

This commit is the enterprise counterpart of odoo/odoo#133196.
The click helper now throws an error if the target is disabled. In
the adapted test, we already assert that the button is disabled,
there's no point clicking on it.

X-original-commit: b0eb7961a60b9786309db4eabd83c4602c7d9cff
Part-of: odoo/enterprise#49004
Co-authored-by: Aaron Bohy <aab@odoo.com>