[FIX] l10n_dk_reports: make formula less restrictive
The aim of this commit is to correct the formula from the reports
allowing users (and Odoo) to create accounts that is not part of the CoA
aside of the existing accounts and make sure those are correctly
included in the reports.

Context:
During the reworking of the CoA we were a bit too fast and we wrote the
accounts numbers writen from the official spreadsheet.

Before this commit:
Accounts created by the OSS feature wouldn't appear in the CoA.

After this commit:
Those accounts appears in the balance sheet

closes odoo/enterprise#48611

Task-id: 3060790
X-original-commit: 1ecbbcdea67cda58c8c8e85931e6d3ad01120ddb
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>