[FIX] pos_iot: use iot_device when printing reports
Current behavior:
After this refactor (https://github.com/odoo/odoo/pull/120070) when you
try to print a report from the POS, the report was always downloaded
even if you set a printer in the report configuration. But the report
should be printed if a printer is set on the report configuration.
It was never taken into account and the report was always downloaded.

Steps to reproduce:
- Install pos_iot, and activate gift card.
- Go in Settings > Reports and look for the gift card report.
- Set a printer on the report configuration.
- Go in the POS and sell a gift card.
- It will be downloaded on the computer, but it should be printed on the
    printer you set on the report configuration.

opw-3495176

closes odoo/enterprise#49317

X-original-commit: a82aef84b1e9f2999fa87500bd7fd43f036c1d99
Related: odoo/odoo#139326
Signed-off-by: Quentin Lejeune (qle) <qle@odoo.com>
Signed-off-by: Robin Engels (roen) <roen@odoo.com>