[FIX] pos_restaurant_appointment: Fix appointment name in floor screen
Previously, an operation was done to retrieve the appointment name. This
operation was needed because when the appointment was created in the front
end, a prefix was added to it. The issue is that this prefix was not added
when the appointment was created in the back end. The solution to remove
the prefix from the appointment name will be done by the team that is
responsible of the Event module.

closes odoo/enterprise#48627

Signed-off-by: Robin Heinz (rhe) <rhe@odoo.com>