[FIX] web_studio: original user service is unaltered
Before this commit, and partly because of odoo/odoo@ba1a5509fa49fd846739252d16083dd8cb334b53,
get_views within studio were made with the studio context key set to a truthy value.

This was because the user context was altered in place when entering studio, meaning that every ORM call from that point had that context key.
In the case of get_views, this is wrong, as only the ViewEditor should load views that way.

For regular ORM calls however, the studio context key is kept for consistency and to avoid
breaking some functional flow.

Also, it makes entering/leaving studio a little safer from a context point of view, as
the user service that studio uses is fully and only contained inside it (the StudioClientAction).

Note that every webClient services that capture the original user service in their closure are unaffected by the studio user service. This is why the get_views works as expected (the viewService
captures the original user service).

Part-of: odoo/enterprise#48640