[FIX] web_studio: do not hang if leaving the formEditor's browser tab
Before this commit, when editing a form view without a record, and trying to close
the browser's tab would warn the user that some changes have occured.

In the studio flow this is irrelevant and even parasitic.

After this commit, there is no warning when closing a tab in which a formEditor is instanciated.

Part-of: odoo/enterprise#48640