[FIX] l10n_de_pos_res_cert: remove string formatting from back-end
Steps to reproduce:
Set up German localization for PoS, make a PoS that is
bar / restaurant, make an order and proceed to pay.

Error pop up: `Unknown Error, Please Contact Odoo Support`

When synchronizing, get traceback:
```
TypeError: (value || 0).toFixed is not a function
```

The issue is that before saas-16.3, unit price was converted to string
and was processed to have correct formating on the Python side.
In saas-16.3, this string was passed into `formatCurrency`
method which did a similar functionality. However `formatCurrency`
expects a number and not a string which causes the issue.

To fix this, it is better to do all the string processing in JavaScript
code.

opw-3499791

closes odoo/enterprise#50918

X-original-commit: 508a91d1f8b58ca4b222d5636a7089ef1cf58b41
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Ahmad Khanalizadeh (khah) <khah@odoo.com>