[IMP] mrp_workorder: shop floor maintain order of the cards by default
Upon interacting with the cards in the shop floor, they would sometimes
"shuffle" around because the data was reloaded and re-sorted from the
backend. As this is a major UX hindrance for the user, this commit
always keeps the ordering the same as long as the user does not press
the reload button or changes the tab. New cards that might appear in the
view will be added at the end of the card list, so the existing ones
also don't shift around.

closes odoo/enterprise#48379

Task: 3456827
Related: odoo/odoo#137535
Signed-off-by: Steve Van Essche <svs@odoo.com>