[FIX] account_reports: Expected date picker in aged reports
Following a conversion of the report to owl, the Expected date picker
seems to not have any effect anymore.
This fix will bring back the expected behaviour (update the date, only for the
report) as it was working before.
As a bonus, it will also pre-populate the datepicker with the date currently set
on the report line.

task id # 3547268

closes odoo/enterprise#48906

X-original-commit: ac6159e510f252e38e562978f9b24aa9b08194f1
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>