[FIX] web_studio: tooltips for fields in the list editor work
Before this commit, and because of odoo/enterprise@025829afaacb8e02bae10e67de7dca0f52fc5d30
hovering on a column in debug mode crashed because the tooltip for fields was wrong,
in fact, the method to get the tooltips info was overriden in the editor without calling super.

After this commit, there is no crash.

closes odoo/enterprise#48640

Signed-off-by: Luca Vitali (luvi) <luvi@odoo.com>