[FIX] web_enterprise, web_studio: fix layout issues
Commit[1] introduced some changes to the studio home menu to fix some
layout and usability.

These changes were applied on Master via a forward-port, but since the
structure of the home studio menu changed, we need to modify a bit the
way the layout and the classes applied to the elements.

Commit[1]: 92bbabd92d320af639c18348e15f5999429e4822

task-3536769

closes odoo/enterprise#48698

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>