[IMP] account_reports: open generated entries as journal items
When viewing the deferred entries in any way, we want to show the move
lines grouped by their move, since that eases auditing on first sight.
This was already changed for most of the actions, apart from the one in
the deferred reports. Also we automatically expand the grouped move
lines now to ease auditing.

task-3525607

Part-of: odoo/enterprise#48834