[IMP] account_reports: allow auditing the deferred reports
Currently it is not very easy to audit the deferred reports.

To improve this, we now allow users to click on any cell of a deferred
report, which takes us to a list view of journal items grouped by
move for the given period and account for that specific cell.

The amounts displayed might defer from the report values, since the
report shows a theoretic computation. This is okay, because when the
user generates entries again, Odoo will compute the difference and the
correct data will be displayed when auditing again.

In general these are the entries displayed for every case:

| Account | Total | Before | Sept 2023 | After |
|---------|-------|--------|-----------|-------|
| 600000  | A     | B      | C         | D     |
| 610000  |       |        |           |       |
| Total   | E     | F      | G         | H     |

- [A]: The vendor bill lines with account 600000 being/to be deferred in
  Sept 2023.
- [B]: The vendor bill lines with account 600000 being/to be deferred in
  Sept 2023 and having an accounting date before Sept 2023, as well as
  their deferral lines with account 600000 falling before Sept 2023.
- [C]: The vendor bill lines with account 600000 being/to be deferred in
  Sept 2023 and having an accounting date in Sept 2023, as well as their
  deferral lines with account 600000 falling in Sept 2023.
- [D]: The deferral lines with account 600000 falling after Sept 2023
  and having an original bill with deferrals also in Sept 2023.
- [E]: Same as [A] but with all expense accounts instead of only 600000.
- [F]: Same as [B] but with all expense accounts instead of only 600000.
- [G]: Same as [C] but with all expense accounts instead of only 600000.
- [H]: Same as [D] but with all expense accounts instead of only 600000.

task-3525607

closes odoo/enterprise#48834

Signed-off-by: John Laterre (jol) <jol@odoo.com>