[FIX] project_enterprise, *: avoid MemoryError in GanttView
When loading up the GanttView from Field of Service
and switching to "Year", the number of tasks to load can be
quite big. If these tasks' description contains images, it's
very likely that a MemoryError will be raised.

To avoid that, turn off field prefetching in _compute_leave_warning
and _gantt_progress_bar_user_ids and then use read to prefetch
all needed fields at once.

Because read (as oppose to fetch in saas-16.2 up) does not check
the cache, the query count in test_performance has to be increased
from 9 to 10, even though creating new tasks inside this test
puts their fields' value in cache.

Browsing for self.ids prevents AccessErrors with NewIds that don't
have values in database yet. But they do in cache so the following
calls to __get__ will simply return the values from the field cache. So
no additionnal queries from turning off prefetching in that case.

opw-3510929

closes odoo/enterprise#51034

X-original-commit: 9e6a60c221b1b322b92a31d2748575cc8c169b36
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>