[IMP] stock_account: country_code in the view
In some localisation (example: latam) we put company dependant field on the
stock.view_picking_form view. In the community PR, we have added the
country_code to the form view in stock_account module so that we can use
directly the country_code, so this commit remove the unnecessary country_code.

closes odoo/enterprise#47043

Task: 3270458
Related: odoo/odoo#134617
Related: odoo/upgrade#5266
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>