[IMP] appointment: allow customizing the appointment mail templates
This commit allows users to customize the mail templates sent when an attendee
books or cancels a meeting through an appointment type.

In addition, it also allows to NOT send one (or both) of those by leaving the
associated field blank.

The goal is to be able to make the template appropriate to the appointment use
case, such as removing the time frame of the appointment for a restaurant
booking, as you don't really want that to appear in the customer email
(e.g: you only have 4 hours to eat).

In addition, we introduce the 'attendee_invitation_mail_template' mail template
inside the data, which is an enhanced copy of the calendar module's
'calendar_template_meeting_invitation', including more information.

Task-3547269

closes odoo/enterprise#48874

Related: odoo/odoo#138547
Related: odoo/upgrade#5269
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>