[FIX] marketing_automation: avoid @extend and fix misalignment
We can remove @extend because it seems that this no longer fix
anything even if the css rule "width: 100%" on .o_field_widget
is still there.

So in this commit, we found an effective fix by simply using w-*
to avoid the fields being one below the other.

Steps to reproduce:
- Go to Marketing Automation
- Click on "New"
- In the form view click on "Add new activity"
- See "Trigger" and "Cancel after" fields

closes odoo/enterprise#48920

Related: odoo/odoo#138615
Signed-off-by: Adrien Dieudonné (adr) <adr@odoo.com>