[IMP] sign: simplify kanban record bottom layout
Actually the `&.o_sign_sticky_bottom` didn't targetted anything as this
class is now a child of the `.o_kanban_record`.

This commit simplifies the layout of the bottom of the card by using
Flex utility classes instead of custom CSS and fixes the slightly broken
layout due to the '&.o_sign_sticky_bottom` targetting.

Note: an SCSS extend was removed in the process.

task-3546717

Part-of: odoo/enterprise#48920