[IMP] pos_preparation_display: add recall button to preparation display
In this PR we had a new button to the navbar of the preparation display.
This new button allows to recall the last order that was sent to the
next stage (if any). It is useful when the order was sent to the next
stage by mistake. This PR also includes some minor changes like the
removal of pos_preparation_display_order_stage records when finishing
the order.

closes odoo/enterprise#48967

Signed-off-by: David Monnom (moda) <moda@odoo.com>