[IMP] industry_fsm, sale_subscription: Ui fix
In this pr: https://github.com/odoo/enterprise/pull/30853 we've merged two
settings into one to make Tax calculation & display more coherent. In the case
of a "Tax per line" computation, the column "Tax Included" is optional=hide in
the backend, but we've elected to put it in the PDF all the time.

This task we will only change the pdf and portal (preview):
1. Remove the "Tax Incl" column on all relevant templates (pdf/portal).
2. Rename the "Tax Excl" column to "Amount".

task-3552656

closes odoo/enterprise#48923

Related: odoo/odoo#138624
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>