[FIX] industry_fsm_sale: prevent removal of delivered Service product
- steps: open fsm app -> select a task -> add products -> select service product -> add some qty -> try to remove qty -> inconsistent behavior and error message.

- Issue: The qty of these products should not be decreasable

- Cause: Service product with the *qty_delivered_method* set as 'manual' are considered as directly delivered so their quantities shouldn't be decreasable once you added quantity.
This logic is applied in the __inverse_fsm_quantity_ method in industry_fsm_sale module

- Fix: adding a min_quantity key in the catalog _sale_product_catalog_update_sale_order_line_info_ method return dictionary,
the value is setted as the sol.qty_delivered. Once this value has been fetched, the kanban_record will include it in the productCatalogData.minimumQuantityOnProduct

I'm open to discussion to implement this in the sale_stock module but i don't think the quantity delivered are updated as soon as
the quantity is added to the sale order anywhere else than in the industry_fsm_stock module.

Task-3553140

closes odoo/enterprise#48940

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>