[IMP] account_reports: improve display of company names in pdf export
Curently the company names are displayed as

Companies: Company 1, Company 2, ....

with the same style (grey) as the rest of the company information directly below.

The "Companies: " prefix is superfluous and dropped in this commit.
The style of the company names is changed (now black) to make it better pop to the
eye compared to the rest of the company information.

closes odoo/enterprise#47735

Related: odoo/odoo#136173
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>