[FIX] website_helpdesk_knowledge: adapt selector in a tour
Beforer this commit, the `button.o_switch_view.oi-view-kanban` no
longer finds the button because of a changes done in community version.
That is, the `oi-view-kanban` class has been moved into a new `i` tag
displayed inside that button.

This commit changes the selector to correcly select the button expected,
that is, the kanban button in views switcher to display the kanban view.

task-3376776

closes odoo/enterprise#48963

Related: odoo/odoo#131162
Related: odoo/upgrade#5272
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>