[FIX] document_spreadsheet: patch browser object correctly
This commit fixes the patch made to the browser object. Before this commit, the
navigator property only has the clipboard property defined, which could cause
other issues, with the code using the browser.navigator object to define some
properties.

It is part of the task-3275275, when the feature detection is made using
browser.navigator.userAgent. A test would fail without this fix.

closes odoo/enterprise#48968

Related: odoo/odoo#135866
Signed-off-by: Pierre Paridans (app) <app@odoo.com>