[FIX] planning : save cancel repeat until recurrence
Issue:
======
When you cancel repeat with type `until` for a `planning slot` it will
not get saved after saving the form.

Steps to reproduce the error:
=============================
- Install planning
- Go to planning create a new slot from the Gantt (default) view.
- Add resource , choose date as a period of 2 days for example (doesn't
  matter), select repeat, change forever to until and select the date
  for some day in next month for example and save it.
- Click on the created slot and unselect repeat and click save.
- Click again on the slot , it will show that repeat is still selected.

Origin of the issue:
====================
The reason behind this issue is that _compute_repeat_until is called
when changing repeat value and thus changing the values of
`slot.repeat_until` to False and then in the save `inverse_repeat` is
called and it will not match all the conditions in the second if : The
condition on `repeat_until` will be False.

opw-3550191

closes odoo/enterprise#50679

X-original-commit: 92006f6db4bb84bba2aa06a815748e89d27ca509
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>