[IMP] iot: convert report <=> printer relation to m2m
Previously only 1 printer could be assigned to a report. We now want to
be able to assign multiple printers to a report so that the user can
choose which printer they want to use. The feature to choose which
printer is being done in a separate PR (see: ), but this change is
needed due to this PR needing to convert printers that are currently
assigned in `stock.picking.type` to the corresponding report record
during the upgrade.

Therefore we do the field conversion + the minimal amount of changes to
iot to ensure the existing printing feature doesn't break (i.e. default
to the first printer to be safe)

Part of task: 3046178
Upgrade PR: odoo/upgrade#5298

Part-of: odoo/enterprise#43362