[FIX] pos*: fix order count and order last preparation changes
pos*: pos_preparation_display, pos_self_order_preparation_display

Before this PR, when an order was made in self with pay after each with
no payment method and service at table. The order that was sent to the
floor plan had its last_order_preparation_change wrongly filled and
the order count sent to the config was also wrong. It is now fixed.

We also fixed the computation of the tracking number and add the feature
that, if a config has only one payment method, it will be automatically
selected when going to the payment screen.

Community PR: odoo/odoo#138813

closes odoo/enterprise#48991

Signed-off-by: David Monnom (moda) <moda@odoo.com>