[FIX] mrp_workorder: Merge context using MakeContext helper
If mrp_workorder_expiry module is installed a skip_expired context
key is defined on the two buttons "Mark as Done" and "Mark as Done
and Close MO". However, the context key was not transmitted properly
to the backend because params.context is a string, whereas
this.props.context is an Object, resulting in a new context were
only the key/values from this.props.context were consistent, while
the string in params.context was exploded in multiple characters.

closes odoo/enterprise#48988

Signed-off-by: William Henrotin (whe) <whe@odoo.com>