[FIX] account_online_synchronization: Fix flickering when passing mouse on widget
When we pass our mouse on the connected until widget
if the text is too big for the place in the div, the widget
flickers. It's not visually pleasant.
We fix the issue by using a text-nowrap class on the div.
NB: To fix completed the flickering, we have to remove
the useless d-flex on the main div. To avoid breaking the stable
policies, we'll fixing it in master only.

closes odoo/enterprise#49046

Task-id: 3556087
X-original-commit: 2c899252da208321cd6f87f1bfd8275373feb7d2
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>