[FIX] l10n_pe_reports: Serie fixed for PE reports
SUNAT, the Peruvian government entity, mandates that document numbers must be transmitted with
all their digits intact. Document types, such as the DUA (unique customs declaration), have a
prefix consisting of 3 alphanumeric digits, while the majority of other document types have a
4-digit prefix. In Odoo, the first letter is consistently generated automatically. The
primary objective of this commit is to ensure that the prefix's digit count is determined
independently of the document type, allowing only the necessary number of digits to be
transmitted in the corresponding column of the report.

Extra, the book number field is optional and Sunat is not validating it properly (Column 20).
Therefore, we will send it empty to avoid inconsistency with the versions of PLE and SIRE.

closes odoo/enterprise#49020

Signed-off-by: Josse Colpaert <jco@odoo.com>