[FIX] account: fix invoice_report_file_id if proforma pdf
When triggering the Send & Print with the allow_fallback_pdf=True, if an
error occurs (e.g.: no UNSCPC code on the product with CFDI), the
proforma PDF is generated instead and is assigned to
`invoice_report_file_id`. However, only a successful PDF should be bound
to this field.

task-3542881

closes odoo/enterprise#49137

Related: odoo/odoo#138206
Signed-off-by: Laurent Smet (las) <las@odoo.com>