[FIX] spreadsheet_dashboard_*,documents_spreadsheet_*: set explicit en_US locale
The en_US default locale in o-spreadsheet is m/d/yyyy but it's mm/dd/yyyy in odoo.
Because of this mismatch, spreadsheets from the source code of odoo (dashboards,
templates) will have a small warning saying the spreadsheet locale and the user
locale mismatch.
That's because existing spreadsheet don't have any locale and the default one is
used.

This commit adds the locale to all o-spreadsheet json files.

Notes:
- we could have completely upgrade the json file format to the latest
  o-spreadsheet version (version: 14), but just adding the settings is enough
- I didn't added the locale to a few template files which are minified (some have
  tens of thousands lines. The diff isn't worth it for such a minor annoyance.

closes odoo/enterprise#49148

X-original-commit: deaa6b8080fa050722fa0762a8f9849f1af4f7ec
Related: odoo/odoo#139133
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>