[FIX] web_studio: report's model can differ from action's
From the reports kanban view, delete the filter on model and click on a report
that has a different model than that of the action.

Before this commit, the loading of the record ids on which the report applies was wrong.

After this commit, the report's model is retrieved from the initial loading of its data.
Hence the model on which it applies is correct.

task-id-3553150

Part-of: odoo/enterprise#49057