[FIX] web_studio: disable the add custom fields in list while in studio
Inside studio, go to the automations tab, switch to list view.

Before this commit the button to "add custom fields" was present.
It doesn't make sense to display it here, so we just don't. Plus, it made the client crash when clicked on.

After this, that button is not there anymore.

task-id-3557774

Part-of: odoo/enterprise#49057