[FIX] account_reports: Gen Ledger avoid unfold all
Issue:
The General Ledger can't open when coming from Profit & Loss by
selecting a specific account to show in the GL.

Analyze:
When coming from another report to the General Ledger with a specific
contextual account while having the `unfold_all` option. The unfold_all
option is passed to the general ledger.
Which leads to a TimeOut issue on databases using a lot accounting
because there are just too many lines.

Fix:
As on V15. When coming from another report and specifying a contextual
account, the report should open with only the data from the specified
account. If the user want to Unfold All it is only 2 more clicks. Which
is either a perf and an ergonomic improvement.

Related ticket:
opw-3452606

Affected version:
16.0 and above

closes odoo/enterprise#49334

X-original-commit: c0c36a2ef4c877f286d3b7eb4dae7739ac5d343c
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>