[IMP] website_sale_renting: disable checkout if cart not ready
Before this commit, a warning on the cart was displayed as soon as the
products in it were not available at the renting dates. But it was still
possible to move on until the payment page, where the customer could not
pay.

This commit makes sure the checkout button is disabled as soon as the
cart is not available at the current renting dates.

task-3537684

closes odoo/enterprise#48533

Related: odoo/odoo#137844
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>