[FIX] pos*: function using field not existing in this module
In the module pos_self_order_preparation_display, we use in the
orders.py controller the field 'self_order_online_payment_method_id'
which is not created in this module nor in any dependent module
but in pos_online_payment. We thus have to move this to
pos_online_payment_self_order_preparation_display to avoid traceback.

Part-of: odoo/enterprise#49200