[FIX] account_online_synchronization: Empty transaction helper not rendering.
Problem
---------
When fetching online transactions, it can happen that no new
transactions are found. In such a case, a nice helper with a message and
an image is meant to be displayed on the screen. However, in some
situations, this helper is not rendered, and only the plain HTML code is
displayed.

Objective
---------
Fix this behavior so that the helper is always rendered.

Solution
---------
After investigating, it appeared that the helper would not be rendered
when the call to `_open_iframe()` is made prior to fetching for new
transactions.

Indeed, this function calls some JS code that performs an ORM call. The
ORM call fetches for the new transactions and returns an
`ir.actions.act_window` in JSON. The returned act_window includes
the helper if no new transactions have been found.
When doAction() is called with act_window as a parameter, the helper
is transformed into a string and is thus not rendered on the screen.

To solve the issue,
1. Make sure that the helper HTML code is safe using escape and Markup.
2. After the `orm.call` and before calling `doAction`, retransform the
helper into a markup explicitely.

Task-3553029

closes odoo/enterprise#49490

X-original-commit: 4e26ab340fc8f7223db19f1a16d9ba8e9661293a
Signed-off-by: Nicolas Viseur (vin) <vin@odoo.com>
Signed-off-by: Antoine Boonen (aboo) <aboo@odoo.com>