[IMP] knowledge: adds save at focusout
This commit adds an autosave feature when the document detects a `mouseleave`
event. This event enables us to only target a specific element without
impacting any of the document's descendants.
This enables Knowledge to save more often in order to fix one of its
biggest issue: the surprise loss of data.

When we detect a `mouseleave` on the document element we call the method
`isDirty` to check if the body has indeed been modified.
If that's the case then we launch the save mechanism to commit the changes
done to the Database.

task-3531468

closes odoo/enterprise#49127

Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>