[FIX] helpdesk_timesheet: make editable unit_amount in form view
Before this commit, the 0613d22 commit makes `unit_amount` field
readonly to not allow the user to edit that field when he goes to
timesheets views via Helpdesk action. The problem is since saas-16.3
version, the form view used in `act_hr_timesheet_line_helpdesk` action
is the same then the one used in `My Timesheets` and so the field
is not only in readonly in `act_hr_timesheet_line_helpdesk` action.

This commit reverts the changes made for that form view to avoid
having that field in readonly in `My Timesheets` menu.

closes odoo/enterprise#49156

X-original-commit: 054c682a0cadefb486a69e4d89b95f2f83cbb17b
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>