[FIX] helpdesk_timesheet: make unit_amount readonly in form view
Before this commit, when the user clicks on the stat button in
helpdesk team to see the timesheets linked to that helpdesk team
the user should not be able to edit `unit_amount` field. Unfortunately,
he can edit that field in form view.

This commit makes that field readonly in the form view of timesheet
only if the user came from the stat button in helpdesk team to
see timesheets linked.

task-3542388

closes odoo/enterprise#48523

Related: odoo/odoo#137820
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>