[IMP] knowledge: allow integration of embedded views and article links in templates
Currently, the templates of Knowledge only contain static HTML with text,
images, tables, lists, etc. Now, we would like to include templates
featuring databases and article links. It will then be possible to
showcase more advanced features of Knowledge.

Unfortunately, these elements refer to an article and require an article
id to work properly. For the article links, for instance, we need to
provide the id of the articles we want to open. When applying a template,
the system will generate new articles with new ids. To ensure that the
links will redirect the user to the right article, we have to update the
ids stored in the document and provide the ids of the newly generated
articles.

As the templates are static and the ids are dynamic, we cannot properly
integrate those elements in a template. This commit aims to eliminate
that limitation by providing a magic `ref` function that people can
use in their templates to refer to an external record. People will then
be able to easily integrate databases and article links in their templates.

This commit will also introduce new templates containing databases and
article links. People can then use those templates to try out the more
advanced features of Knowledge and understand the power of Knowledge
with real world examples.

Technical aspects:

To streamline our code base, we will merge the model `knowledge.article`
with `knowledge.article.template`. That way, the articles and the templates
will always be consistent. By merging the two models, we will not have to
duplicate the logic and the views for the template preview. It will also
not be necessary to: transform the favorites set on the embedded views,
update the default values set in the context, etc. when turning a template
into an article.

During the development, we considered different approaches to tackle the
existing limitations of the static templates:

1. We considered making the ids optional by introducing fallback mechanisms.
   The idea was to use the id of the active record when no id was provided.
   This would have allowed the integration of embedded view of article items
   without providing any id. However, the approach was very limited: it was
   not possible to list the article items of another article or insert an
   article link.
2. We considered patching the generated template with `lxml` to update the
   ids stored in the document. The solution didn't work very well: In some
   cases, it was not possible to know if the ids should be transformed
   or not as the ids could refer to another model.
3. We considered turning the static templates into dynamic templates using
   `QWeb`. The solution was checking all the boxes but we preferred not
   to allow non-admin users creating templates and using `QWeb` because
   the rendering engine relies on the `safe_eval` function which can
   potentially have undiscovered security flaws. To not take any risk,
   we prefer not to use that approach.

For all these reasons, we prefer including a magic `ref` function in the
templates and post-process the template to replace the `ref` function
call with the appropriate id. With that approach, people can explicitly
tell which id should be dynamically transformed.

task-3334717

closes odoo/enterprise#48691

Related: odoo/upgrade#5258
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>